requirements.txt
# Webflow Article GA4 Traffic Tracker
# Install with: pip install -r requirements.txt

python-dotenv>=1.0.0
requests>=2.28.0

# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.23.0
gspread>=5.12.0

# Data handling
pandas>=2.0.0

# Web app (for cloud trigger)
flask>=3.0.0
gunicorn>=21.0.0
